views merged into POST

+++ core/apps/notes/views.py
@@ -10,8 +10,23 @@ from ..users.models import Users
 
 
 class NotesManagement(APIView):
+    """Base API class for notes management
+
+    HTTP methods allowed:
+        GET, POST
+    """
+
     def get(self, request, format=None):
-        status_message = {"response": ""}
+        """GET method
+
+        HTTP method allowed:
+            GET
+
+        Args:
+            request: First argument Django http base class provide.
+            format: Body format.
+        """
+        status_message = {"response": "No notes found"}
         status_http = status.HTTP_404_NOT_FOUND
 
         user_id = request.data.get("id")
@@ -33,55 +48,57 @@ class NotesManagement(APIView):
         return Response(status_message, status_http)
 
     def post(self, request, format=None):
-        status_message = {"response": "post method"}
+        """POST method
+
+        HTTP method allowed
+            POST, PUT, DELETE
+
+        Args:
+            request: First argument Django http base class provide.
+            format: Body format.
+        """
+        status_message = {"response": "generic post method"}
         status_http = status.HTTP_406_NOT_ACCEPTABLE
 
-        user_info = NotesSerializer(request.data)
-        query = Users.objects.filter(pk=user_info.data.get("id")).first()
+        action = request.headers.get("action")
 
-        if query:
-            new_note = Notes.objects.create(
-                title=user_info.data.get("title"),
-                description=user_info.data.get("description"),
-                user_id=query,
-            )
-            new_note.save()
+        if action == "POST":
+            user_info = NotesSerializer(request.data)
+            query = Users.objects.get(pk=user_info.data.get("id"))
 
-            status_message = {"response": {"id": new_note.id}}
-            status_http = status.HTTP_201_CREATED
+            if query:
+                new_note = Notes.objects.create(
+                    title=user_info.data.get("title"),
+                    description=user_info.data.get("description"),
+                    user_id=query,
+                )
+                new_note.save()
 
-        return Response(status_message, status_http)
+                status_message = {"response": {"id": new_note.id}}
+                status_http = status.HTTP_201_CREATED
 
-    def delete(self, request, format=None):
-        status_message = {"response": "delete method"}
-        status_http = status.HTTP_404_NOT_FOUND
+        if action == "DELETE":
+            id_note = request.data.get("id")
+            query = Notes.objects.get(pk=id_note)
 
-        id_note = request.data.get("id")
-        query = Notes.objects.get(pk=id_note)
+            if query:
+                query.delete(keep_parents=True)
 
-        if query:
-            query.delete(keep_parents=True)
+                status_message = {"response": {"message": "Note deleted"}}
+                status_http = status.HTTP_200_OK
 
-            status_message = {"response": {"message": "Note deleted"}}
-            status_http = status.HTTP_200_OK
+        if action == "PUT":
+            note_info = NotesSerializer(request.data)
+            id_note = note_info.data.get("id")
+            query = Notes.objects.get(pk=id_note)
 
-        return Response(status_message, status_http)
+            if query:
+                query.title = str(note_info.data.get("title"))
+                query.description = str(note_info.data.get("description"))
 
-    def put(self, request, format=None):
-        status_message = {"response": "put method"}
-        status_http = status.HTTP_404_NOT_FOUND
+                query.save()
 
-        note_info = NotesSerializer(request.data)
-        id_note = note_info.data.get("id")
-        query = Notes.objects.get(pk=id_note)
-
-        if query:
-            query.title = str(note_info.data.get("title"))
-            query.description = str(note_info.data.get("description"))
-
-            query.save()
-
-            status_message = {"response": {"message": "Note modified"}}
-            status_http = status.HTTP_200_OK
+                status_message = {"response": {"message": "Note modified"}}
+                status_http = status.HTTP_200_OK
 
         return Response(status_message, status_http)
